server/pyproject.toml
[project]
name = "template-server"
version = "1.0.0"
description = "Template server"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.2",
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "python-socketio>=5.11.4",
    "requests>=2.32.3",
    "uvicorn>=0.31.1",
]
